improve the method for obtaining the message.

+++ QuasarFireOperation.Domain/CommandModel/Requests/AddSatelites/AddSatelliteObject.cs
@@ -7,7 +7,7 @@ namespace QuasarFireOperation.Domain.CommandModel.Requests.AddSatelites
     public class AddSatelliteObject
     {
         public string Name { get; set; }
-        public float Distance { get; set; }
+        public double Distance { get; set; }
         public List<string> Message { get; set; }
     }
 }
